Add files via upload
This is my react-app which is give u a functionality in changing text from upper case to lower, lower to upper and more function like that, it does't matter how much the length of paragraph just took your para and put it in it they will replace the all para like you want. It also gives the length of character and time taken to read it even they have a dark mode to readuced the light effect which is harmful for you eyes,
just enjoying the functionality of my react thank you.